pyproject.toml
[tool.poetry]
name = "starter"
version = "0.1.0"
description = "Fast API Starter project clean architecture, repository pattern"
authors = ["Aurthur Musendame <aurthurmusendame@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.0"
uvicorn = {extras = ["all"], version = "^0.21.1"}
fastapi = "^0.95.1"
SQLAlchemy = "^2.0.10"
asyncpg = "^0.27.0"
opentelemetry-instrumentation-fastapi = "^0.38b0"
opentelemetry-sdk = "^1.17.0"
opentelemetry-api = "^1.17.0"
opentelemetry-exporter-otlp = "^1.17.0"
opentelemetry-instrumentation-logging = "^0.38b0"
opentelemetry-instrumentation-sqlalchemy = "^0.38b0"
pytz = "^2023.3"
pydantic = {extras = ["email"], version = "^1.10.7"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.172.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
mypy = "^1.2.0"
isort = "^5.12.0"
black = "^23.3.0"
autoflake = "^2.1.1"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
